refact: adjust exports.
